+++ drift_dev/lib/src/analysis/results/table.dart
@@ -53,6 +53,8 @@ class DriftTable extends DriftElementWithResultSet {
   /// `customConstraints` getter in the table class with this value.
   final List<String> overrideTableConstraints;
 
+  DriftColumn? _rowIdColumn;
+
   DriftTable(
     super.id,
     super.declaration, {
@@ -69,7 +71,18 @@ class DriftTable extends DriftElementWithResultSet {
     this.virtualTableData,
     this.writeDefaultConstraints = true,
     this.overrideTableConstraints = const [],
-  });
+  }) {
+    _rowIdColumn = DriftColumn(
+      sqlType: DriftSqlType.int,
+      nullable: false,
+      nameInSql: 'rowid',
+      nameInDart: 'rowid',
+      declaration: declaration,
+      isImplicitRowId: true,
+    )..owner = this;
+  }
+
+  late final DriftColumn? rowid = _findRowId();
 
   /// Whether this is a virtual table, created with a `CREATE VIRTUAL TABLE`
   /// statement in SQL.
@@ -93,6 +106,25 @@ class DriftTable extends DriftElementWithResultSet {
         .toSet();
   }
 
+  DriftColumn? _findRowId() {
+    if (withoutRowId) return null;
+
+    // See if we have an integer primary key as defined by
+    // https://www.sqlite.org/lang_createtable.html#rowid
+    final primaryKey = fullPrimaryKey;
+    if (primaryKey.length == 1) {
+      final column = primaryKey.single;
+      if (column.sqlType == DriftSqlType.int ||
+          column.sqlType == DriftSqlType.bigInt) {
+        // So this column is an alias for the rowid
+        return column;
+      }
+    }
+
+    // Otherwise, expose the implicit rowid column.
+    return _rowIdColumn;
+  }
+
   /// Determines whether [column] would be required for inserts performed via
   /// companions.
   bool isColumnRequiredForInsert(DriftColumn column) {
@@ -106,15 +138,14 @@ class DriftTable extends DriftElementWithResultSet {
       return false;
     }
 
-    // A column isn't required if it's an alias for the rowid, as explained
-    // at https://www.sqlite.org/lang_createtable.html#rowid
-    final fullPk = fullPrimaryKey;
-    final isAliasForRowId = !withoutRowId &&
-        column.sqlType == DriftSqlType.int &&
-        fullPk.length == 1 &&
-        fullPk.single == column;
+    if (rowid == column) {
+      // If the column is an alias for the rowid, it will get set automatically
+      // by sqlite and isn't required for inserts either.
+      return false;
+    }
 
-    return !isAliasForRowId;
+    // In other cases, we need a value for inserts into the table.
+    return true;
   }
 
   @override