+++ examples/with_built_value/lib/database.g.dart
@@ -32,7 +32,10 @@ class _$Foo extends Foo {
 
   @override
   int get hashCode {
-    return $jf($jc(0, moorField.hashCode));
+    var _$hash = 0;
+    _$hash = $jc(_$hash, moorField.hashCode);
+    _$hash = $jf(_$hash);
+    return _$hash;
   }
 
   @override
@@ -84,4 +87,4 @@ class FooBuilder implements Builder<Foo, FooBuilder> {
   }
 }
 
-// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,deprecated_member_use_from_same_package,lines_longer_than_80_chars,no_leading_underscores_for_local_identifiers,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new,unnecessary_lambdas
+// ignore_for_file: deprecated_member_use_from_same_package,type=lint